DP-1299 : fix failing tests
The behavior of std::filesystem::equivalent is different between gcc
(rhel8) which does not throw an exception when only one of the paths does
not exist and clang 15 on Mac OSX which does throw.  For files to be
copied into working directories, a check is made that the files exist.
If so, a subsequent check is made to ensure that files are not the same
as the destination directory.  This second check was throwing exceptions
on Mac OSX but behaving as intended on gcc.

The fix is to use a nothrow version of std::filesystem::equivalent.