Resolving print output error for NonDLHSSampling
When variance_based_decomp was specified, output would only print
Sobol' indices, ignoring any other requested statistics, such as
response moments or cumulative distributions. This is desired for
pick-freeze because of the non-random structure of the samples, but
it's undesireable for the binned method which operates on random
samples. This behavior is now corrected.