Fix hanging test on windows
Becuase direct execution is currently broken on windows, when
the script is run this way with the system interface, it receives
no CL args. It aborts but Dakota doesn't know anything is wrong
and it waits for a results file until the jenkins job times out.

Set the results file to 'results.out' and make the driver write
FAIL to it when it aborts.