mfmc_updates: generalize NonDNonHierarchSampling::scale_to_budget_with_pilot()
MFMC solutions with model reordering on the fly are showing dependence on
initial model ordering.  This function was one of the main offenders as it
enforced r_i > 1 using the user-given ordering.  Now it follows evaluation
ratio-ordering.

Some additional review of initial group costs in bound constraints is still
needed, e.g. for the finite bounds used by DIRECT.